requirements.txt
cloudinary==1.29.0
GitPython==3.1.27
PyGithub==1.55
python-dotenv==0.20.0
requests==2.27.1
telegram==0.0.1
python-telegram-bot==20.0a2